created password validation as middleware and updated api code
